daemon/requirements.txt
certifi==2019.9.11
chardet==3.0.4
falcon==2.0.0
idna==2.8
requests==2.22.0
urllib3==1.25.6
PyYAML==5.1.2
uwsgi